improvements on management commands

+++ web/public/management/commands/clean_and_seed.py
@@ -4,32 +4,46 @@ from public.management.seed import initial_data
 from public import models
 
 class Command(BaseCommand):
-    def handle(self, *args, **options):
-        data = args[0] if args else initial_data.data
+    def add_arguments(self, parser):
+        parser.add_argument('model', type=str, nargs='?', default=None)
 
-        if not args:
+    def handle(self, *args, **options):
+        model = options.get('model', None)
+        if model:
+            data = initial_data.by_model[model]
+            print(data)
             self._clean_and_seed(data)
         else:
+            data = initial_data.data
             self._clean_and_seed(data)
 
-        self._check_if_seeded(data)
+            related_data = args[1] if args else initial_data.related
+            related_data = [f() for f in related_data]
+            self._clean_and_seed(related_data)
+
+        print('Seed completed...')
 
     def _clean(self, data):
         for documents in data:
-            type(documents[0]).objects.all().delete()
+            if len(documents) > 0:
+                print(documents)
+                print('cleaning ', documents[0].__class__)
+                type(documents[0]).objects.all().delete()
 
 
     def _seed(self, data):
         for models in data:
-            for model in models:
-                model.save()
+            if len(models) > 0:
+                print(models)
+                print('seeding ', models[0].__class__)
+                for model in models:
+                    model.save()
 
     def _clean_and_seed(self, data):
         self._clean(data)
         self._seed(data)
-        print (models.Chart.objects.all())
 
-    def _check_if_seeded(self, data):
+    '''def _check_if_seeded(self, data, data_type):
         counts = map(
             lambda models: (
                 self._get_seeded_models_count(models), len(models)
@@ -38,8 +52,8 @@ class Command(BaseCommand):
 
         different_counts = list(filter(lambda x: x[0] != x[1], counts))
 
-        print("Seed successful...") if not different_counts else print("Seed failed...")
-
+        print("Seed of ", data_type, " successful...") if not different_counts else print("Seed failed...")
+    '''
     def _get_seeded_models_count(self, objects):
         model = type(objects[0])
         return model.objects.all().count()
